// content/content-source/contentful/types.d.ts
export interface IContactDetailsFields {
  /** Label */
  label: string;

  /** Value */
  value: string;

  /** href */
  href?: string | undefined;
}

/** A single entry for the contact detaiils */

export interface IContactDetails extends Entry<IContactDetailsFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "contactDetails";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IContentDetailsFields {
  /** Title */
  title: string;

  /** Contact Details */
  contactDetails?: IContactDetails[] | undefined;
}

/** A collection of contact details */

export interface IContentDetails extends Entry<IContentDetailsFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "contentDetails";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ICvFields {
  /** Title */
  title: string;

  /** Personal Statement */
  personalStatement: IPersonalStatement;

  /** Skills */
  skills: ISkills;

  /** Work History */
  workHistory: IWorkHistory;

  /** Contact Details */
  contactDetails: IContentDetails;
}

/** Curriculum Vitae */

export interface ICv extends Entry<ICvFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "cv";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPersonalStatementFields {
  /** Title */
  title: string;

  /** Statement */
  statement: string;
}

/** A summary of the CV */

export interface IPersonalStatement extends Entry<IPersonalStatementFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "personalStatement";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IRoleDescriptionFields {
  /** Buisness Name */
  buisnessName: string;

  /** Job Title */
  jobTitle: string;

  /** Start Date */
  startDate: string;

  /** endDate */
  endDate?: string | undefined;

  /** Description */
  description: string;
}

/** Summary of a role */

export interface IRoleDescription extends Entry<IRoleDescriptionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "roleDescription";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ISkillListFields {
  /** Title */
  title: string;

  /** Skills */
  skills: string[];
}

/** A list of skills */

export interface ISkillList extends Entry<ISkillListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "skillList";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ISkillsFields {
  /** Section Title */
  groupTitle: string;

  /** Description */
  description?: string | undefined;

  /** Skills Lists */
  skillsLists: ISkillList[];
}

/** A list of lists of skills */

export interface ISkills extends Entry<ISkillsFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "skills";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IWorkHistoryFields {
  /** Title */
  title: string;

  /** Roles */
  roles: IRoleDescription[];
}

/** History of roles */

export interface IWorkHistory extends Entry<IWorkHistoryFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "workHistory";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}